commission subscription

+++ app/Http/Controllers/AgentController.php
@@ -6,6 +6,7 @@ use Carbon\Carbon;
 use App\Models\Property;
 use App\Models\Commission;
 use App\Models\Membership;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -15,36 +16,45 @@ class AgentController extends Controller
         $members = Membership::where('member_id', auth()->id())->first();
         $present_dateTime = Carbon::now()->format('Y-m-d H:i:s');
 
-        if($present_dateTime > $members->membership_date){
-            $members->delete();
+        if($members){
+            if($present_dateTime > $members->membership_date){
+                $members->delete();
+            }
         }
         return view('backend.agent.dashboard');
     }
 
     public function property_index()
     {
-        $properties = Property::latest()->get();
+        $properties = Property::where('status',1)->get();
         return view('backend.agent.property.index',compact('properties'));
     }
 
-    public function request_owned(Request $request)
+    public function send_request(Request $request)
     {   
-       $property = Property::findOrFail($request->property_id);
+        $property = Property::findOrFail($request->property_id);
 
-        $property->agent_status = $request->status;
+        $property->agent_status = 2;
+        $property->agent_id = Auth::id();
 
         $property->save();
 
-        // Commission 
-        Commission::create([
-            'user_id' => Auth::id(),
-            'price' => $property->price,
-            'percent' => 9,
-            'admin_fee' => round($property->price * 9 / 100),
+        return response()->json([
+            'message' => "Request Sent"
         ]);
+    }
+
+    public function cancel_request(Request $request)
+    {   
+        $property = Property::findOrFail($request->property_id);
+
+        $property->agent_status = NULL;
+        $property->agent_id = NULL;
+
+        $property->save();
 
         return response()->json([
-            'message' => "Status Updated"
+            'message' => "Request Cancelled"
         ]);
     }
 }
